+++ tw-tkms-starter/src/main/java/com/transferwise/kafka/tkms/dao/TkmsPostgresDao.java
@@ -8,4 +8,9 @@ public class TkmsPostgresDao extends TkmsDao {
   protected String getInsertSql(TkmsShardPartition shardPartition) {
     return "insert into " + getTableName(shardPartition) + " (message) values (?) returning id";
   }
+  
+  @Override
+  protected String getSelectSql(TkmsShardPartition shardPartition) {
+    return "select id, message from " + getTableName(shardPartition) + " order by id limit ?";
+  }
 }